Fixed percentages. Sometime between the end of June and beginning of July Woot! changed from a decimal representation of percent sold (0.00-1.00) to an integer representation (0-100).

+++ library/Ds/Service/WootFeed.php
@@ -215,7 +215,7 @@ class Ds_Service_WootFeed {
         $item['teaser'] = (string)$wootNamespace->teaser;
         $history['comments'] = (int)$wootNamespace->comments;
         $history['sold_out'] = (strtolower((string)$wootNamespace->soldout) == 'true') ? true : false;
-        $history['percent_sold'] = (float)$wootNamespace->soldoutpercentage;
+        $history['percent_sold'] = (int)$wootNamespace->soldoutpercentage;
         $history['updated'] = date('c');
         $images['standard'] = (string)$wootNamespace->standardimage;
         $images['detail'] = (string)$wootNamespace->detailimage;
